some minor API changes/improvements
git-svn-id: http://svn.code.sf.net/p/nitro-nitf/code/trunk@812 7dc3ea37-dbd3-4c13-bb9a-7ccd57e75167

+++ c/nitf/include/nitf/Types.h
@@ -101,7 +101,7 @@ typedef int8_t             nitf_Int8;
 typedef int16_t            nitf_Int16;
 typedef int32_t            nitf_Int32;
 typedef int64_t            nitf_Int64;
-typedef int                nitf_IOHandle;
+typedef long               nitf_IOHandle;
 typedef off_t              nitf_Off;
 
 typedef void*             NITF_DLL_FUNCTION_PTR;
